Refresh with new Demo application
The demo application has been replaced with a WebAssembly version that
can run in the browser. The previous demo was very hard to build and a
significant barrier. The new demo will be easier to build, run and adapt.

Besides the demo there are some boyscout cleanups in the code base:
* Added LICENSE file
* Extended .gitignore file
* Updated README.md
* Reordered the code in rsmotion.cpp to fix some warnings